Move WTP into the packages where it is supposed to be.
