+++ lib-extra/src/wtp/java/com/diffplug/spotless/extra/glue/wtp/html/StructuredDocumentProcessor.java
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package com.diffplug.spotless.extra.eclipse.wtp.html;
+package com.diffplug.spotless.extra.glue.wtp.html;
 
 import java.util.ArrayList;
 import java.util.Arrays;
