+++ lib-extra/src/wtp/java/com/diffplug/spotless/extra/glue/wtp/sse/CleanupStep.java
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package com.diffplug.spotless.extra.eclipse.wtp.sse;
+package com.diffplug.spotless.extra.glue.wtp.sse;
 
 import org.eclipse.wst.sse.core.internal.cleanup.AbstractStructuredCleanupProcessor;
